+++ Include/Goblin/User/Interface/AnnounceDialog.hpp
@@ -17,7 +17,7 @@
 
 #include <libtcod.hpp>
 
-#include "Goblin/User/Interface/Drawable.hpp"
+#include <Goblin/User/Interface/Abstract/Drawable.hpp>
 #include "Goblin/User/Interface/Dialog.hpp"
 
 class AnnounceDialog : public Scrollable
