+++ Include/Goblin/User/Interface/SquadsDialog.hpp
@@ -25,7 +25,7 @@
 #include "Goblin/User/Interface/Frame.hpp"
 #include "Goblin/User/Interface/Dialog.hpp"
 #include "Goblin/User/Interface/UIList.hpp"
-#include "Goblin/User/Interface/Drawable.hpp"
+#include <Goblin/User/Interface/Abstract/Drawable.hpp>
 
 class SquadsDialog : public Dialog
 {
